refactor: add FontRenderer that manages shaders for font rendering
