+++ Cargo.toml
@@ -37,6 +37,7 @@ percent-encoding = "2.2.0"
 bytemuck = "1.13.1"
 font-kit = "0.14.3"
 pathfinder_geometry = "0.5.1"
+pathfinder_color = "0.5.0"
 png = "0.17.16"
 
 [build-dependencies]
